requirements.txt
flask>=3.0
pandas>=2.0
openpyxl>=3.1
watchdog>=4.0
apscheduler>=3.10
xlrd>=2.0
pdfplumber>=0.10
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
